airflow-astro/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
astronomer-cosmos==1.0.4
apache-airflow-providers-postgres==5.6.0
apache-airflow-providers-dbt-cloud
astronomer-cosmos
apache-airflow-providers-airbyte

